Handle entity path in event hub connection string
# Conflicts:
#	x-pack/filebeat/input/azureeventhub/config.go
